move shared observeSaveComplete functionality into CreationFragment
